fix(frontend): component splitted
