+++ packages/frontend/components/Shared/ConfirmTransactionModal.tsx
@@ -1,21 +1,8 @@
 import CloseIcon from '@mui/icons-material/Close';
-import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined';
-import {
-  Box,
-  Button,
-  Card,
-  Dialog,
-  Divider,
-  Paper,
-  Stack,
-  Tooltip,
-  Typography,
-} from '@mui/material';
+import { Box, Button, Dialog, Paper, Stack, Typography } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
-import { RoutingStep, RoutingStepDetails, Token } from '@x-fuji/sdk';
-import { formatUnits } from 'ethers/lib/utils';
+import { RoutingStepDetails } from '@x-fuji/sdk';
 import Image from 'next/image';
-import { ReactNode } from 'react';
 
 import {
   ActionType,
@@ -23,12 +10,12 @@ import {
   recommendedLTV,
   remainingBorrowLimit,
 } from '../../helpers/assets';
-import { chainName } from '../../helpers/chains';
 import { BasePosition } from '../../helpers/positions';
-import { camelize, formatValue, toNotSoFixed } from '../../helpers/values';
+import { formatValue } from '../../helpers/values';
 import { FetchStatus } from '../../store/borrow.store';
-import { NetworkIcon } from './Icons';
-import TokenIcon from './Icons/TokenIcon';
+import AssetBox from './ConfirmationTransaction/AssetBox';
+import InfoRow from './ConfirmationTransaction/InfoRow';
+import RouteBox from './ConfirmationTransaction/RouteBox';
 import WarningInfo from './WarningInfo';
 
 type ConfirmTransactionModalProps = {
@@ -311,315 +298,6 @@ export function ConfirmTransactionModal({
         </Button>
       </Paper>
     </Dialog>
-  );
-}
-
-function AssetBox({
-  type,
-  token,
-  value,
-  isEditing,
-  actionType,
-}: {
-  type: 'debt' | 'collateral';
-  token: Token;
-  value: string;
-  isEditing: boolean;
-  actionType: ActionType;
-}) {
-  const { palette } = useTheme();
-  const labelMap =
-    isEditing && actionType === ActionType.REMOVE
-      ? { debt: 'Payback', collateral: 'Withdraw' }
-      : { debt: 'Borrow', collateral: 'Deposit' };
-
-  const label = labelMap[type];
-
-  const descriptionMap =
-    isEditing && actionType === ActionType.REMOVE
-      ? {
-          debt: 'The network where the debt will be paid back',
-          collateral: 'The network from where the collateral will be withdrawn',
-        }
-      : {
-          debt: 'The network where the loan will be taken out',
-          collateral: 'The network where the collateral will be deposited on',
-        };
-
-  const description = descriptionMap[type];
-
-  return (
-    <Card
-      variant="outlined"
-      sx={{
-        borderColor: palette.secondary.light,
-        m: '0.5rem 0',
-        width: '100%',
-        order:
-          isEditing && actionType === ActionType.REMOVE && type === 'debt'
-            ? 1
-            : 2,
-      }}
-    >
-      <Stack
-        width="100%"
-        flexDirection="row"
-        alignItems="center"
-        justifyContent="space-between"
-      >
-        <Typography variant="small">{label}</Typography>
-
-        <Stack flexDirection="row" alignItems="center" gap={0.75}>
-          <TokenIcon token={token} height={16} width={16} />
-          <Typography variant="small">
-            {`${formatValue(value, {
-              maximumFractionDigits: 3,
-            })} ${token.symbol}`}
-          </Typography>
-        </Stack>
-      </Stack>
-
-      <Divider sx={{ m: '0.75rem 0', height: '1px', width: '100%' }} />
-
-      <Stack
-        width="100%"
-        flexDirection="row"
-        alignItems="center"
-        justifyContent="space-between"
-      >
-        <Stack flexDirection="row" alignItems="center">
-          <Typography variant="small">Network</Typography>
-
-          {!isEditing && (
-            <Tooltip
-              title="If you wish to open a position on a different chain, please select an alternative route."
-              placement="top"
-            >
-              <InfoOutlinedIcon
-                sx={{
-                  ml: '0.313rem',
-                  fontSize: '0.875rem',
-                  display: { xs: 'none', sm: 'inline' },
-                  color: '#919191',
-                  cursor: 'pointer',
-                }}
-              />
-            </Tooltip>
-          )}
-        </Stack>
-
-        <Stack flexDirection="row" alignItems="center" gap={0.75}>
-          <NetworkIcon network={token.chainId} height={16} width={16} />
-          <Typography variant="small">{chainName(token.chainId)}</Typography>
-        </Stack>
-      </Stack>
-
-      <Typography
-        textAlign="start"
-        mt=".5rem"
-        variant="xsmall"
-        fontSize="0.75rem"
-        lineHeight="1.2rem"
-        color={palette.info.main}
-        sx={{ width: '100%' }}
-      >
-        {description}
-      </Typography>
-    </Card>
-  );
-}
-
-function InfoRow({ title, value }: { title: string; value: ReactNode }) {
-  const { palette } = useTheme();
-
-  return (
-    <Stack
-      width="100%"
-      flexDirection="row"
-      alignItems="center"
-      justifyContent="space-between"
-      mt="0.6rem"
-      flexWrap="wrap"
-    >
-      <Typography
-        sx={{
-          ['@media screen and (max-width: 370px)']: {
-            fontSize: '0.7rem',
-          },
-        }}
-        color={palette.info.main}
-        variant="small"
-      >
-        {title}
-      </Typography>
-
-      {value}
-    </Stack>
-  );
-}
-
-function RouteBox({
-  steps,
-  isCrossChainTransaction,
-}: {
-  steps: RoutingStepDetails[];
-  isCrossChainTransaction: boolean;
-}) {
-  const { palette } = useTheme();
-
-  const stepsToShow = steps.filter(
-    (s) => s.step !== RoutingStep.START && s.step !== RoutingStep.END
-  );
-
-  function description(step: RoutingStepDetails) {
-    if (step.lendingProvider) {
-      const withToPrepositions = [RoutingStep.DEPOSIT, RoutingStep.PAYBACK];
-      const preposition = withToPrepositions.includes(step.step)
-        ? 'to'
-        : 'from';
-      return (
-        <Typography
-          align="center"
-          variant="xsmall"
-          fontSize="0.75rem"
-          sx={{ display: 'flex', gap: '0.25rem' }}
-        >
-          {`${preposition} ${step.lendingProvider.name} on `}
-          <NetworkIcon
-            network={chainName(step.chainId)}
-            height={14}
-            width={14}
-          />
-        </Typography>
-      );
-    }
-    if (step.step === RoutingStep.X_TRANSFER) {
-      return (
-        <Stack flexDirection="row" alignItems="center" gap="0.25rem">
-          <Typography align="center" variant="xsmall" fontSize="0.75rem">
-            from
-          </Typography>
-          <NetworkIcon
-            network={chainName(step.token?.chainId)}
-            height={14}
-            width={14}
-          />
-          <Typography align="center" variant="xsmall" fontSize="0.75rem">
-            to
-          </Typography>
-          <NetworkIcon
-            network={chainName(step.chainId)}
-            height={14}
-            width={14}
-          />
-        </Stack>
-      );
-    }
-
-    return <></>;
-  }
-
-  function textForStep({ step, amount, token }: RoutingStepDetails) {
-    switch (step) {
-      case RoutingStep.DEPOSIT:
-      case RoutingStep.BORROW:
-      case RoutingStep.PAYBACK:
-      case RoutingStep.WITHDRAW:
-        return camelize(
-          `${step.toString()} ${toNotSoFixed(
-            formatUnits(amount ?? 0, token?.decimals || 18)
-          )} ${token?.symbol}`
-        );
-      case RoutingStep.X_TRANSFER:
-        return camelize(
-          `${step.toString()} ${toNotSoFixed(
-            formatUnits(amount ?? 0, token?.decimals || 18)
-          )} ${token?.symbol}`
-        );
-      default:
-        return camelize(step);
-    }
-  }
-
-  return (
-    <Card
-      variant="outlined"
-      sx={{
-        borderColor: palette.secondary.light,
-        m: '0.5rem 0 1rem 0',
-        width: '100%',
-      }}
-    >
-      <Stack
-        width="100%"
-        flexDirection="row"
-        alignItems="center"
-        justifyContent="space-between"
-      >
-        <Typography variant="small">Route</Typography>
-
-        {isCrossChainTransaction && (
-          <Typography variant="small">via Connext Network</Typography>
-        )}
-      </Stack>
-
-      <Divider sx={{ m: '0.75rem 0', height: '1px', width: '100%' }} />
-
-      <Stack
-        width="100%"
-        justifyContent="space-between"
-        sx={{
-          flex: 1,
-          maxWidth: '30rem',
-          flexWrap: 'wrap',
-          flexDirection: { xs: 'column', sm: 'row' },
-          alignItems: 'center',
-          justifyContent: { xs: 'center', sm: 'space-between' },
-          gap: '0.5rem',
-        }}
-      >
-        {stepsToShow.map((step, i) => (
-          <>
-            {i !== 0 && (
-              <Box
-                key={i}
-                sx={{
-                  ['@media screen and (max-width: 600px)']: {
-                    transform: 'rotate(90deg)',
-                  },
-                }}
-              >
-                <Image
-                  alt="Arrow icon"
-                  src="/assets/images/shared/doubleArrow.svg"
-                  height={10}
-                  width={9}
-                />
-              </Box>
-            )}
-            <Stack
-              key={step.step}
-              direction="column"
-              sx={{
-                p: '0.375rem 0.45rem',
-                backgroundColor: '#35353B',
-                borderRadius: '6px',
-                minWidth: '8rem',
-                flex: 1,
-                width: { xs: '100%', sm: 'unset' },
-              }}
-            >
-              <Typography align="left" variant="xsmall">
-                {textForStep(step)}
-              </Typography>
-              <Typography align="left" variant="xsmall" mt={0.5}>
-                {description(step)}
-              </Typography>
-            </Stack>
-          </>
-        ))}
-      </Stack>
-    </Card>
   );
 }
